♻️ extract trackLocationChanges
